+++ rust-src/concordium_base/src/protocol_level_tokens/token_module_initialization_parameters.rs
@@ -1,7 +1,7 @@
-use concordium_base_derive::{CborDeserialize, CborSerialize};
-
 use super::{MetadataUrl, TokenAmount};
-use crate::protocol_level_tokens::token_holder::CborTokenHolder;
+use crate::{common::cbor::value, protocol_level_tokens::token_holder::CborTokenHolder};
+use concordium_base_derive::{CborDeserialize, CborSerialize};
+use std::collections::HashMap;
 
 /// These parameters are passed to the token module to initialize the token.
 /// The token initialization update will also include the ticker symbol,
@@ -25,6 +25,9 @@ pub struct TokenModuleInitializationParameters {
     pub mintable:           Option<bool>,
     /// Whether the token is burnable.
     pub burnable:           Option<bool>,
+    /// Additional fields.
+    #[cbor(other)]
+    pub additional:         HashMap<String, value::Value>,
 }
 
 #[cfg(test)]
@@ -57,6 +60,7 @@ mod test {
             initial_supply:     Some(TokenAmount::from_raw(10000000, 8)),
             mintable:           Some(true),
             burnable:           Some(true),
+            additional:         Default::default(),
         };
 
         let cbor = cbor::cbor_encode(&token_module_initialization_parameters).unwrap();